Fix and tidy patch.c rs_patch_s_copying() to use rs_long_t for length.
Fix s_patch_s_copying() to use rs_long_t for the request copy length, and make
it -Wconversion compliant. This would have broken long file support on 32bit
platforms.

Improve trace output from rs_patch_s_cmdbyte(). Make rs_patch_s_params() more
-Wconversion compliant by using size_t for a buffer length.

Change prototab.h len_1 and len_2 fields from size_t to int, which makes
various code more -Wconversion compliant (netint.c use int for lengths of
these parameters).

+++ src/prototab.h
@@ -34,7 +34,7 @@
 typedef struct rs_prototab_ent {
     enum rs_op_kind kind;
     int immediate;
-    size_t len_1, len_2;
+    int len_1, len_2;
 } rs_prototab_ent_t;
 
 extern const rs_prototab_ent_t rs_prototab[];
